// components/Form/Form.types.ts
export interface IFormInputs {
	data: string;
	case_number: string;

	claimant_name: string;
	claimant_registration: string;
	claimant_address: string;
	claimant_email: string;
	claimant_phone_code: string;
	claimant_phone_number: string;

	debtor_name: string;
	debtor_registration: string;

	confirm_information: boolean;
}